requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.43.0
Django==5.2
djangorestframework==3.16.0
dotenv==0.9.9
python-dotenv==1.1.0
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
